app/backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
redis
python-dotenv
pydantic
pydantic-settings
sqlalchemy
alembic
fastapi-mail
